pyproject.toml
[project]
name = "facebook-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "chromadb>=1.3.5",
    "click==8.1.8",
    "fastapi[standard]>=0.121.3",
    "flask==3.1.0",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.24.2",
    "google-api-python-client==2.167.0",
    "google-auth==2.39.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.2",
    "google-genai==1.11.0",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.71.0",
    "grpcio-status==1.71.0",
    "gspread==6.2.0",
    "gunicorn==23.0.0",
    "h11==0.14.0",
    "httpcore==1.0.8",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "oauthlib==3.2.2",
    "packaging==25.0",
    "proto-plus==1.26.1",
    "protobuf==5.29.4",
    "psutil>=7.1.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "pyparsing==3.2.3",
    "python-dotenv==1.1.0",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "tqdm==4.67.1",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "uritemplate==4.1.1",
    "urllib3==2.4.0",
    "websockets==15.0.1",
    "werkzeug==3.1.3",
]
